@@ -129,70 +126,69 @@
     "native-is-elevated": "^0.4.3",
     "ngx-google-analytics": "^11.2.1",
     "request": "^2.88.2",
-    "rxjs": "^7.5.7",
+    "rxjs": "~7.5.0",
     "tslib": "^2.3.1",
     "universal-analytics": "^0.5.3",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^13.0.0",
-    "@angular-devkit/build-angular": "^13.1.4",
-    "@angular-devkit/core": "^13.1.4",
-    "@angular-devkit/schematics": "^13.1.4",
-    "@angular/animations": "13.1.3",
-    "@angular/cdk": "13.1.3",
-    "@angular/cli": "13.1.4",
-    "@angular/common": "13.1.3",
-    "@angular/compiler": "13.1.3",
-    "@angular/compiler-cli": "13.1.3",
-    "@angular/core": "13.1.3",
+    "@angular-builders/custom-webpack": "^14.0.1",
+    "@angular-devkit/build-angular": "14.2.6",
+    "@angular-devkit/core": "14.2.6",
+    "@angular-devkit/schematics": "14.2.6",
+    "@angular/animations": "14.2.6",
+    "@angular/cdk": "14.2.5",
+    "@angular/cli": "~14.2.0",
+    "@angular/common": "14.2.6",
+    "@angular/compiler": "14.2.6",
+    "@angular/compiler-cli": "14.2.6",
+    "@angular/core": "14.2.6",
     "@angular/fire": "^7.4.1",
-    "@angular/flex-layout": "^13.0.0-beta.36",
-    "@angular/forms": "13.1.3",
-    "@angular/language-service": "13.1.3",
-    "@angular/platform-browser": "13.1.3",
-    "@angular/platform-browser-dynamic": "13.1.3",
-    "@angular/platform-server": "13.1.3",
-    "@angular/router": "13.1.3",
+    "@angular/flex-layout": "^14.0.0-beta.41",
+    "@angular/forms": "14.2.6",
+    "@angular/language-service": "14.2.6",
+    "@angular/platform-browser": "14.2.6",
+    "@angular/platform-browser-dynamic": "14.2.6",
+    "@angular/platform-server": "14.2.6",
+    "@angular/router": "14.2.6",
     "@ffxiv-teamcraft/simulator": "^3.1.4",
     "@ffxiv-teamcraft/simulator-cn": "npm:@ffxiv-teamcraft/simulator@3.0.14",
     "@ffxiv-teamcraft/simulator-kr": "npm:@ffxiv-teamcraft/simulator@3.0.14",
-    "@firebase/messaging": "^0.9.6",
+    "@firebase/messaging": "^0.9.19",
     "@fortawesome/fontawesome": "^1.1.8",
     "@fortawesome/fontawesome-free-brands": "^5.0.13",
     "@fortawesome/fontawesome-free-solid": "^5.0.13",
     "@fortawesome/fontawesome-free-webfonts": "^1.0.9",
     "@kaiu/ng-serializer": "1.1.1",
     "@kaiu/serializer": "1.4.0",
-    "@ngrx/effects": "13.0.2",
-    "@ngrx/entity": "13.0.2",
-    "@ngrx/router-store": "13.0.2",
-    "@ngrx/schematics": "13.0.2",
-    "@ngrx/store": "13.0.2",
-    "@ngrx/store-devtools": "13.0.2",
-    "@nguniversal/builders": "^12.1.3",
-    "@nguniversal/express-engine": "^12.1.3",
+    "@ngrx/effects": "14.0.2",
+    "@ngrx/entity": "14.0.2",
+    "@ngrx/router-store": "14.0.2",
+    "@ngrx/schematics": "14.0.2",
+    "@ngrx/store": "14.0.2",
+    "@ngrx/store-devtools": "14.0.2",
+    "@nguniversal/builders": "^14.2.0",
+    "@nguniversal/express-engine": "^14.2.0",
     "@nguniversal/module-map-ngfactory-loader": "^8.2.6",
     "@ngx-translate/core": "^14.0.0",
-    "@ngx-translate/http-loader": "^6.0.0",
-    "@nrwl/angular": "13.4.6",
-    "@nrwl/cypress": "13.4.6",
-    "@nrwl/workspace": "13.4.6",
+    "@ngx-translate/http-loader": "^7.0.0",
+    "@nrwl/angular": "14.8.4",
+    "@nrwl/cypress": "14.8.4",
+    "@nrwl/workspace": "14.8.4",
     "@swimlane/dragula": "^3.8.0",
-    "@swimlane/ngx-charts": "^20.0.1",
+    "@swimlane/ngx-charts": "^20.1.0",
     "@swimlane/ngx-dnd": "^9.0.0",
     "@types/express": "^4.17.0",
     "@types/express-serve-static-core": "^4.17.13",
-    "@types/jasmine": "3.6.11",
+    "@types/jasmine": "4.0.3",
     "@types/jasminewd2": "~2.0.3",
     "@types/jwt-decode": "^2.2.1",
     "@types/lodash": "^4.14.178",
-    "@types/node": "14.14.45",
-    "@types/request": "^2.48.5",
-    "@types/semver": "^5.4.0",
+    "@types/node": "16.11.26",
+    "@types/request": "^2.48.8",
+    "@types/semver": "^7.3.12",
     "@xivapi/angular-client": "1.8.1",
-    "algoliasearch": "^4.11.0",
-    "angular-feather": "^5.3.0",
+    "algoliasearch": "^4.14.2",
     "angular2-chartjs": "^0.5.1",
     "apollo-angular": "^1.8.0",
     "apollo-angular-link-http": "^1.9.0",
@@ -219,62 +216,63 @@
     "del-cli": "^1.1.0",
     "dexie": "^3.2.0",
     "dotenv": "10.0.0",
-    "electron": "^18.2.4",
-    "electron-builder": "^23.0.3",
-    "electron-builder-squirrel-windows": "^23.0.9",
+    "electron": "^21.1.1",
+    "electron-builder": "^23.6.0",
+    "electron-builder-squirrel-windows": "^23.6.0",
     "enquirer": "^2.3.6",
     "express": "^4.17.1",
-    "fast-json-patch": "^3.0.0-1",
+    "fast-json-patch": "^3.1.1",
     "file-saver": "^2.0.5",
     "firebase": "^9.11.0",
-    "firebase-tools": "^10.0.1",
-    "graphql": "^14.5.0",
-    "graphql-tag": "^2.10.0",
+    "firebase-tools": "^11.14.2",
+    "graphql": "^16.6.0",
+    "graphql-tag": "^2.12.6",
     "hammerjs": "^2.0.8",
     "husky": "^0.14.3",
     "inquirer-autocomplete-prompt": "^1.0.1",
     "jasmine": "2.99.0",
-    "jasmine-core": "~2.99.1",
+    "jasmine-core": "4.2.0",
     "jasmine-marbles": "0.8.4",
-    "jasmine-spec-reporter": "~4.2.1",
+    "jasmine-spec-reporter": "7.0.0",
     "jsdom": "^18.0.0",
     "jsonlint-cli": "^1.0.1",
-    "karma": "~6.3.4",
+    "karma": "6.4.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
+    "karma-coverage": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "5.1.0",
+    "karma-jasmine-html-reporter": "2.0.0",
     "less": "^4.1.2",
-    "memoizee": "^0.4.14",
+    "memoizee": "^0.4.15",
     "multi-progress": "^2.0.0",
     "ng-drag-drop": "^5.0.0",
     "ng-push-ivy": "^1.0.8",
-    "ng-zorro-antd": "13.2.2",
-    "ngx-color-picker": "^11.0.0",
-    "ngx-markdown": "^13.1.0",
+    "ng-zorro-antd": "14.1.0",
+    "ngx-color-picker": "^13.0.0",
+    "ngx-markdown": "^14.0.1",
+    "nx": "14.8.4",
     "nz-tslint-rules": "^0.901.2",
-    "p-queue": "^7.1.0",
-    "papaparse": "^5.2.0",
-    "prettier": "2.3.1",
-    "quicktype-core": "^6.0.71",
-    "semver": "^5.4.1",
+    "papaparse": "^5.3.2",
+    "prettier": "2.6.2",
+    "quicktype-core": "^6.1.0",
+    "semver": "^7.3.8",
     "standard-version": "^4.4.0",
     "subscriptions-transport-ws": "^0.9.19",
-    "terser-webpack-plugin": "^1.2.3",
-    "ts-loader": "^9.2.6",
-    "ts-node": "9.1.1",
+    "terser-webpack-plugin": "^5.3.6",
+    "ts-loader": "^9.4.1",
+    "ts-node": "10.9.1",
     "tslint": "6.1.3",
-    "typescript": "4.5.5",
+    "typescript": "4.8.4",
     "utf-8-validate": "^5.0.7",
     "validate-commit-msg": "^2.14.0",
     "ws": "^8.2.3",
     "xhr2": "^0.2.1",
     "yargs": "^17.1.1",
-    "zone.js": "0.11.5"
+    "zone.js": "0.11.8"
   },
   "resolutions": {
-    "@types/node": "14.14.45"
+    "@types/node": "16.11.26"
   },
   "config": {
     "commitizen": {
