@@ -1,6 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { NzModalRef } from 'ng-zorro-antd/modal';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 
 @Component({
   selector: 'app-text-question-popup',
@@ -15,7 +15,7 @@ export class TextQuestionPopupComponent implements OnInit {
   @Input()
   placeholder = '';
 
-  public control: FormControl;
+  public control: UntypedFormControl;
 
   constructor(private modalRef: NzModalRef) {
   }
@@ -25,7 +25,7 @@ export class TextQuestionPopupComponent implements OnInit {
   }
 
   ngOnInit(): void {
-    this.control = new FormControl(this.baseText);
+    this.control = new UntypedFormControl(this.baseText);
   }
 
 }
