@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup } from '@angular/forms';
 import { GearsetsFacade } from '../../../modules/gearsets/+state/gearsets.facade';
 import { distinctUntilChanged, expand, filter, first, last, map, shareReplay, switchMap, switchMapTo, takeUntil, tap } from 'rxjs/operators';
 import { TeamcraftComponent } from '../../../core/component/teamcraft-component';
@@ -38,7 +38,7 @@ export class GearsetEditorComponent extends TeamcraftComponent implements OnInit
 
   public machinaToggle = false;
 
-  itemFiltersform: FormGroup = this.fb.group({
+  itemFiltersform: UntypedFormGroup = this.fb.group({
     ilvlMin: [540],
     ilvlMax: [999],
     elvlMin: [1],
@@ -334,7 +334,7 @@ export class GearsetEditorComponent extends TeamcraftComponent implements OnInit
     })
   );
 
-  constructor(private fb: FormBuilder, private gearsetsFacade: GearsetsFacade,
+  constructor(private fb: UntypedFormBuilder, private gearsetsFacade: GearsetsFacade,
               private activatedRoute: ActivatedRoute, private xivapi: XivapiService,
               private lazyData: LazyDataFacade, private cd: ChangeDetectorRef,
               public translate: TranslateService, private dialog: NzModalService,
