@@ -244,8 +244,8 @@ export abstract class FirestoreStorage<T extends DataModel> extends DataStore<T>
     this.recordOperation('write', rows.map(row => row.$key));
     const batch = writeBatch(this.firestore);
     rows.forEach(row => {
-      const ref = this.docRef(row.$key) as DocumentReference<R>;
-      return batch.update<R>(ref, { index: row.index } as UpdateData<R>);
+      const ref = this.docRef(row.$key) as DocumentReference<T & {index: number}>;
+      return batch.update<T & {index: number}>(ref, { index: row.index } as UpdateData<T & {index: number}>);
     });
     return from(batch.commit());
   }
