@@ -1,25 +1,21 @@
-module.exports = function() {
+module.exports = function () {
   return {
     basePath: '',
     frameworks: ['jasmine', '@angular-devkit/build-angular'],
-    plugins: [
-      require('karma-jasmine'),
-      require('karma-chrome-launcher'),
-      require('karma-coverage-istanbul-reporter'),
-      require('@angular-devkit/build-angular/plugins/karma')
-    ],
-    coverageIstanbulReporter: {
-      reports: ['html', 'lcovonly', 'text-summary'],
+    plugins: [require('karma-jasmine'), require('karma-chrome-launcher'), require('karma-coverage'), require('@angular-devkit/build-angular/plugins/karma')],
+    coverageReporter: {
+      subdir: '.',
+      reporters: [{ type: 'html' }, { type: 'text-summary' }],
       dir: './coverage',
       skipFilesWithNoCoverage: true,
       'report-config': {
         html: {
-          subdir: 'html'
-        }
-      }
+          subdir: 'html',
+        },
+      },
     },
     client: {
-      clearContext: false // leave Jasmine Spec Runner output visible in browser
+      clearContext: false, // leave Jasmine Spec Runner output visible in browser
     },
     port: 9876,
     browserNoActivityTimeout: 30000,
@@ -29,9 +25,9 @@ module.exports = function() {
     customLaunchers: {
       ChromeHeadlessNoSandbox: {
         base: 'ChromeHeadless',
-        flags: ['--no-sandbox']
-      }
+        flags: ['--no-sandbox'],
+      },
     },
-    singleRun: false
+    singleRun: false,
   };
 };
